增加长轮询和监听器

+++ api/config-api/dto/request/long_polling_request.go
@@ -7,7 +7,7 @@ type LongPollingRequest struct {
 
 // ConfigKeyVersion 配置键及其版本
 type ConfigKeyVersion struct {
-	NamespaceID int64  `json:"namespace_id" binding:"required,min=1"` // 命名空间ID
+	NamespaceID int    `json:"namespace_id" binding:"required,min=1"` // 命名空间ID
 	ConfigKey   string `json:"config_key" binding:"required"`         // 配置键
 	Version     string `json:"version" binding:"required"`            // 当前客户端持有的版本号（MD5）
 }
