+++ api/config-api/dto/vo/long_polling_vo.go
@@ -9,7 +9,7 @@ type LongPollingResponse struct {
 
 // ConfigChangeDetail 配置变更详情
 type ConfigChangeDetail struct {
-	NamespaceID int64  `json:"namespace_id"` // 命名空间ID
+	NamespaceID int    `json:"namespace_id"` // 命名空间ID
 	ConfigKey   string `json:"config_key"`   // 配置键
 	Version     string `json:"version"`      // 最新版本号（MD5）
 	Value       string `json:"value"`        // 配置值
